+++ components/Modal/index.ts
@@ -1,4 +1,0 @@
-import ModalWrapper from "./wrapper";
-
-export type { IModalPropTypes } from "./types";
-export default ModalWrapper;