+++ CHANGE_LOG.md
@@ -4,9 +4,19 @@
 
 > 2023-04-19
 
+### Fixed
+
+- fixed: Fixing the placeholder error in the settings page input box and addressing the issue with missing language configurations.
+- fixed: when change the channel and the list do not scroll to bottom
+
 ### Add
 
 - feat: Introducing Sentry to perform error log detection
+
+### Changed
+
+- feat: Introducing @radix-ui/react-dialog to enhance the interactive effect of the Modal Component
+- feat: Change the rule of NEXT_PUBLIC_OPENAI_API_KEY and NEXT_PUBLIC_OPENAI_API_PROXY
 
 ## v0.0.3
 
@@ -21,7 +31,7 @@
 - feat: Support for modifying conversation titles.
 - feat: Added the function to clear all conversations.
 - feat: Support for deleting and undoing conversation content.
-- feat: Introducing Radix-UI dialog to enhance the interactive effect of the confirm event
+- feat: Introducing @radix-ui/react-alert-dialog to enhance the interactive effect of the confirm event
 
 ## v0.0.2